+++ Source/OpenAL/oalCaptureDevice.h
@@ -128,7 +128,7 @@ public:
             erase(it);
     }
 	
-    UInt32 Size () const { return size(); }
+    UInt32 Size () const { return UInt32(size()); }
     bool Empty () const { return empty(); }
 };
